Initial set up of versioning and releases (#36)
* chore(releases): initial set up of versioning and releases

* docs(releases): document releases in CONTRIBUTING.md

* chore(releases): fix change set config

Should deal with message:

Some errors occurred when validating the changesets config:
The package or glob expression "oxa-monorepo" is specified in the `ignore` option but it is not found in the project. You may have misspelled the package name or provided an invalid glob expression. Note that glob expressions must be defined according to https://www.npmjs.com/package/micromatch.